New feature: Size task in list according to priority (font size)
Same thing as writing urgent tasks bigger on a board or paper.

This is a optional feature. There's a new choice in the settings.

Translators, can you please update your translation?
Please take a look at the diff of file languages/taskfreak-fr_FR.po
It's only one string: "Size task in list according to priority".
Strings for Yes / No are already translated.

This will be released as version 1.1.0